P4/requirements.txt
tensorflow
keras
pandas
scikit-learn
numpy
liac-arff
